python_template/requirements.txt
argparse
colorama
requests
validator